graph, graphql: Move Value and ValueType into graph::data::store
